requirements.txt
streamlit>=1.31
requests>=2.28
plotly>=5.14
pandas>=2.0
tqdm>=4.65
python-dotenv>=1.0